+++ src/main/java/com/NetCracker/Exceptions/DoctorNotFoundException.java
@@ -1,7 +1,7 @@
 package com.NetCracker.Exceptions;
 
 public class DoctorNotFoundException extends RuntimeException {
-    public DoctorNotFoundException(Integer id) {
+    public DoctorNotFoundException(Long id) {
         super("Could not find a doctor " + id);
     }
 }
