+++ src/main/java/com/NetCracker/Repositories/AppointmentRegistrationRepo.java
@@ -9,12 +9,13 @@ import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 
-public interface AppointmentRegistrationRepo extends JpaRepository<AppointmentRegistration, Integer> {
+public interface AppointmentRegistrationRepo extends JpaRepository<AppointmentRegistration, Long> {
+    //        @Query("select a from AppointmentRegistration a where a.doctor.lastName = 'ФроловВВ' order by a.id asc")
     @Query("select a from AppointmentRegistration a order by a.id asc")
     public List<AppointmentRegistration> findAllByOrderByIdAsc();
 
     @Query("select a from AppointmentRegistration a where a.id = :id")
-    Optional<AppointmentRegistration> findById(@Param("id") Integer id);
+    Optional<AppointmentRegistration> findById(@Param("id") Long id);
 
     @Query("select a from AppointmentRegistration a, Doctor d, Patient p where a.doctor.id = d.id and a.patient.id = p.id and p.lastName = :pat and d.lastName = :doc and a.start = (select min(aa.start) from AppointmentRegistration aa, Doctor dd, Patient pp where aa.doctor.id = dd.id and aa.patient.id = pp.id and pp.lastName = :pat and dd.lastName = :doc)")
     Optional<AppointmentRegistration> findByDoctorAndPatient(@Param("doc") String doc, @Param("pat") String pat);
