+++ src/main/java/com/NetCracker/Repositories/AppointmentRepo.java
@@ -10,12 +10,13 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface AppointmentRepo extends JpaRepository<Appointment, Integer> {
+public interface AppointmentRepo extends JpaRepository<Appointment, Long> {
+    //    @Query("select a from Appointment a where a.appointmentRegistration.doctor.lastName = 'ФроловВВ' order by a.id asc")
     @Query("select a from Appointment a order by a.id asc")
     List<Appointment> findAllByOrderByIdAsc();
 
     @Query("select a from Appointment a where a.id = :id")
-    Optional<Appointment> findById(@Param("id") Integer id);
+    Optional<Appointment> findById(@Param("id") Long id);
 
 }
