+++ src/main/java/com/NetCracker/Repositories/PatientRepository.java
@@ -9,12 +9,12 @@ import org.springframework.data.repository.query.Param;
 import java.util.List;
 import java.util.Optional;
 
-public interface PatientRepository extends JpaRepository<Patient, Integer> {
+public interface PatientRepository extends JpaRepository<Patient, Long> {
     @Query("select a from Patient a order by a.id asc")
     public List<Patient> findAllByOrderByIdAsc();
 
     @Query("select a from Patient a where a.id = :id")
-    Optional<Patient> findById(@Param("id") Integer id);
+    Optional<Patient> findById(@Param("id") Long id);
 
     @Query("select a from Patient a where a.lastName = :pat")
     Optional<Patient> findPatientByLastName(@Param("pat") String pat);
